requirements.txt
# Core Data Science Libraries (Python 3.13 compatible)
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep Learning (use tensorflow-cpu for easier install)
tensorflow>=2.15.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Dashboard & Web
streamlit>=1.29.0

# Data Generation & Utilities
faker>=22.0.0
python-dateutil>=2.8.2
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.1
